chore: Split code up and add linting.
