+++ .jshintrc
@@ -1,0 +1,48 @@
+{
+    "bitwise"       : false,
+    "curly"         : false,
+    "eqeqeq"        : true,
+    "forin"         : true,
+    "immed"         : true,
+    "latedef"       : true,
+    "newcap"        : false,
+    "noarg"         : true,
+    "noempty"       : true,
+    "nonew"         : true,
+    "undef"         : true,
+    "trailing"      : true,
+    "unused"        : "vars",
+    "indent"        : 2,
+    "maxdepth"      : 5,
+    "maxlen"        : 80,
+    "asi"           : false,
+    "boss"          : true,
+    "debug"         : false,
+    "eqnull"        : false,
+    "esnext"        : false,
+    "evil"          : false,
+    "expr"          : true,
+    "funcscope"     : false,
+    "globalstrict"  : false,
+    "iterator"      : false,
+    "lastsemic"     : false,
+    "laxbreak"      : true,
+    "laxcomma"      : false,
+    "loopfunc"      : false,
+    "multistr"      : false,
+    "onecase"       : false,
+    "proto"         : false,
+    "regexdash"     : false,
+    "scripturl"     : false,
+    "smarttabs"     : false,
+    "shadow"        : false,
+    "sub"           : false,
+    "supernew"      : false,
+    "validthis"     : true,
+    "browser"       : false,
+    "node"          : true,
+    "devel"         : true,
+    "jquery"        : false,
+    "white"         : true,
+    "quotmark"      : "single"
+}
