+++ .travis.yml
@@ -4,4 +4,5 @@ node_js:
   - "0.11"
 
 script:
+  - npm run lint
   - npm test
