+++ lib/defaults.js
@@ -1,0 +1,11 @@
+
+// Default options
+module.exports = {
+  ignoreCase: false,
+  literal: false,
+  wordRegexp: false,
+  context: 2,
+  color: {
+    lineno: 'yellow'
+  }
+};
